requirements.txt
fastapi
uvicorn
openai>=1.0.0
sqlalchemy
python-dotenv
httpx
psycopg2-binary
python-multipart